ReadinessKeeper added
